New ball deflection algorithm incl. new Ball prefab
The new movement angle of the ball (upon colliding with a paddle) is now calculated based on the collision position only (without considering the ball's initial direction). Furthermore, bounciness is now handled via script instead of a physics material, which is more predictable, and based on the public SpeedIncrement field (set from the Inspector).